+++ src/JShark/Console.hs
@@ -3,14 +3,15 @@
   , OverloadedStrings
 #-}
 -- | Wrappers over the JS @console@ object.
+-- Import qualified; names clash with 'Prelude'.
 module JShark.Console
   ( log
   , warn
-  , error_
+  , error
   , info
   ) where
 
-import Prelude hiding (log)
+import Prelude hiding (error, log)
 import JShark.Api
 import JShark.Rec ((<:), Rec(..))
 import JShark.Types
@@ -21,8 +22,8 @@ log x = toSyntax_ (ffi "console.log" (arg x <: RecNil))
 warn :: Expr f u -> EffectSyntax f ()
 warn x = toSyntax_ (ffi "console.warn" (arg x <: RecNil))
 
-error_ :: Expr f u -> EffectSyntax f ()
-error_ x = toSyntax_ (ffi "console.error" (arg x <: RecNil))
+error :: Expr f u -> EffectSyntax f ()
+error x = toSyntax_ (ffi "console.error" (arg x <: RecNil))
 
 info :: Expr f u -> EffectSyntax f ()
 info x = toSyntax_ (ffi "console.info" (arg x <: RecNil))